Tracking (global spacing) center
Robofont window to compare several fonts or instances of a font. It
behaves similarly to the Space Center (except glyph spacing can’t be
modified), you can type in text that is displayed simultaneously in all
the fonts displayed (up to 6).
You can change the typeface displayed on each line at any time, and
apply tracking modification, both global (on all fonts) or local (each
font of each line). The tracking shown in this window is a preview, the
font isn’t directly modified. There are [Apply] buttons if you want to
implement this tracking on the fonts.