Better handling of step references
At each filtering step, the resulting glyph is stored so it can be
referenced later in the process if need be. It used to be that this
storing happened before booleanOperations were performed, which led to
occasionally storing an incomplete image of the processed glyph. +
still need to figure out a way to handle components with filter
functions and pens not based on BasePen.